Gets the auth working in the client
All authentication forms are working and behaving as expected. I
took the code that was in index.js for the template modal and moved
it into a new folder within the scripts called 'pages' and named the
file page-events. then loaded the click handler from that file when
the document is ready in the index.js just like we do for auth.
next is to hide modals when successful operations are happening.