Hide/show bg image and dashboard
Added code/css to hide and show the background image with text
and dashboard based on whether or not a user is authenticated.
Upon signin, background image with welcome text disappears and
dashboard appears.  Upon signout, dashboard disappears and background
image appears.
Commented out the @media queries because they are not functioning.  We
can revisit and remove code or rework if able to do so.
All landing page info, including b/g image, is now included in the
"landing-page-content" div so it's easy to show/hide.  All dashboard
content will live in the "dashboard" div.